SimplePaymentsDialog: simplify clearing the form values on dialog close
When the dialog is being closed, we want to clear the form fields, so that on the next open,
the user doesn't see a "flash of previous form content", where the old values, already saved and
not wanted any more, would be visible for a brief moment.

This patch moves the clearing logic to `componentWillReceiveProps` and removes it from all
other places that do close the dialog and had to handle the clearing correctly.